tests/useAuth.test.tsx
import { renderHook, act } from '@testing-library/react-hooks';
import useAuth from '../hooks/useAuth';

// Mock the wagmi and next-auth/react hooks
jest.mock('wagmi', () => ({
  useAccount: jest.fn(),
  useContractRead: jest.fn(),
  useNetwork: jest.fn(),
}));

jest.mock('next-auth/react', () => ({
  useSession: jest.fn(),
}));

describe('useAuth Hook', () => {
  it('returns admin authentication status', async () => {
    // Set up the mock return values for hooks
    require('wagmi').useAccount.mockReturnValue({ address: '0xAdminAddress' });
    require('wagmi').useContractRead.mockReturnValue({ data: '0xAdminAddress' });
    require('wagmi').useNetwork.mockReturnValue({ chain: { id: 5 } });
    require('next-auth/react').useSession.mockReturnValue({ data: { user: { name: 'Mock User' } }, status: 'authenticated' });

    const { result, waitForNextUpdate } = renderHook(() => useAuth());

    await waitForNextUpdate();

    expect(result.current.isAdminAuthenticated).toBe(true);
    expect(result.current.isUserAuthenticated).toBe(true);
    expect(result.current.isLoading).toBe(false);
  });

  // Additional tests...
});
